refactor: moved customise event-and module panels to their component files
